+++ admin/lang/en_US.php
@@ -657,4 +657,5 @@ $i18n = array(
 "CLICK_TO_USE_URL" => "Click to use this URL",
 "ENABLE_CODE_ED" => "<b>Enable the Code editor</b>",
 "PAGE_SUBTITLE" => "Page Subtitle",
+"PAGE_SUMMARY" => "Page Summary",
 );